refactor engine_factory
